+++ student/student.go
@@ -1,0 +1,9 @@
+package student
+import (
+	"gorm.io/gorm"
+)
+
+
+type Student struct {
+	DB *gorm.DB
+}
